Edit: UI: hide overflow-x scrollbar on wide output
The purpose of `overflow-x: scroll` was to deal with a div quirk
that stretched it to the right and not aligning with other items
even if there was no reason to (all children had `max-width 100%`)

And `overflow-x: hidden` achieves the same thing
without the scrollbar.
Which is why it is used here.