Refactor: Use Interface not ABC for controlllers
... where ABC is shorthand for Abstract Base Classes.

Originally thought requring `this.update = this.update.bind(this)`
seemed pretty tedious for (potential) maintainers of this system.

But it turns out React does this,
and there isn't much boilerplate anyway,
at least not enough to use an abstract class that does magic.

You can see an implementation with arrow functions in `Controllers`