Edit: Set colorscheme to use same colors
Originally we assigned colors to the appearing subperiods
and overwrote p1 and pFull cells' colors.
(pFull cells being the cells oscillating at full period)

This was convenient but meant the colorscheme changed
depending on whether there were p1 cells
(as the color assigned would be masked)

Currently we assign colors to the subperiods except p1 and pFull,
and as a result the colorscheme is the same
as long as the same number of colors are used.